More core cleanup
