+++ core/memory/index.ts
@@ -2,17 +2,7 @@
 // https://docs.google.com/spreadsheets/d/17xrEzJk5-sCB9J2mMJcVnzhbE-XH_NvczVSQH9OHvRk/edit?usp=sharing
 // Public exports for our memory "class"
 
-export {
-  Memory,
-  initializeCartridge,
-  setPixelOnFrame,
-  getRgbPixelStart,
-  setLeftAndRightOutputForAudioQueue,
-  storePaletteByteInWasmMemory,
-  loadPaletteByteFromWasmMemory,
-  getSaveStateMemoryOffset,
-  loadFromVramBank
-} from './memory';
+export { Memory, initializeCartridge, getSaveStateMemoryOffset } from './memory';
 
 export {
   eightBitLoadFromGBMemoryWithTraps,
